fix: improve type safety and consistency across authentication system
